Spring Data Jpa has been added to the project, Coach Repository, Service, Controller have been written for new technology.

+++ src/main/java/by/podvintsev/martialartshallsproject/entity/Coach.java
@@ -15,7 +15,7 @@ import lombok.NoArgsConstructor;
 public class Coach {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id_coach;
+    private Long id_coach;
     private String first_name;
     private String last_name;
     private String address;
